+++ frontend/javascripts/oxalis/model_initialization.js
@@ -212,15 +212,35 @@ function maybeWarnAboutUnsupportedLayers(layers: Array<APIDataLayer>): void {
   }
 }
 
-function initializeTracing(annotation: APIAnnotation, tracing: HybridServerTracing) {
+function initializeTracing(_annotation: APIAnnotation, tracing: HybridServerTracing) {
   // This method is not called for the View mode
   const { dataset } = Store.getState();
+  let annotation = _annotation;
 
   const { allowedModes, preferredMode } = determineAllowedModes(dataset, annotation.settings);
   _.extend(annotation.settings, { allowedModes, preferredMode });
 
   const { controlMode } = Store.getState().temporaryConfiguration;
   if (controlMode === ControlModeEnum.TRACE) {
+    if (Utils.getUrlParamValue("sandbox")) {
+      annotation = {
+        ...annotation,
+        restrictions: {
+          ...annotation.restrictions,
+          allowUpdate: true,
+          allowSave: false,
+        },
+      };
+    } else {
+      annotation = {
+        ...annotation,
+        restrictions: {
+          ...annotation.restrictions,
+          allowSave: annotation.restrictions.allowUpdate,
+        },
+      };
+    }
+
     Store.dispatch(initializeAnnotationAction(annotation));
 
     serverTracingAsVolumeTracingMaybe(tracing).map(volumeTracing => {